# Introduce Temple attribute compiler