--- a/haml.gemspec
+++ b/haml.gemspec
@@ -19,7 +19,7 @@ Gem::Specification.new do |spec|
 
   spec.required_ruby_version = '>= 2.0.0'
 
-  spec.add_dependency 'temple', '0.7.7'
+  spec.add_dependency 'temple', '>= 0.8.0'
   spec.add_dependency 'tilt'
 
   spec.add_development_dependency 'rails', '>= 4.0.0'