--- a/lib/haml/temple_engine.rb
+++ b/lib/haml/temple_engine.rb
@@ -1,4 +1,5 @@
 require 'temple'
+require 'haml/escapable'
 require 'haml/generator'
 
 module Haml
@@ -28,6 +29,8 @@ module Haml
 
     use :Parser,   -> { options[:parser_class] }
     use :Compiler, -> { options[:compiler_class] }
+    use Escapable
+    filter :ControlFlow
     filter :MultiFlattener
     filter :StaticMerger
     use Generator