… test/temple_line_counter_test.rb