## Changes
- Implement attribute compiler using Temple
  - To compile attributes, introduce attribute parser using Ripper imprted from Hamlit.
     - On Ripper-less platforms, attribute compiler isn't applied but it fallbacks to current implementation.
  - For almost all cases, we can render attributes without `Haml::Buffer#attributes`, which is a bottleneck of Haml 4.
- Apply Temple's static analyzer to render attributes beforehand

All tests passed without changing behaviors in this branch. They are the last changes I want to release in Haml 5.

## Benchmark
With Ruby 2.4.0 and  [k0kubun/haml_bench/templates/slim_bench.haml](https://github.com/k0kubun/haml_bench/blob/8d8cc944bdc1138c58629109b6ecc06e7d5ea1a7/templates/slim_bench.haml),

### before (current Haml's master)
[the result on Travis](https://travis-ci.org/k0kubun/haml_bench/jobs/199977841)

```
Rendering: /home/travis/build/k0kubun/haml_bench/templates/slim_bench.haml
Warming up --------------------------------------
          haml 4.0.7     1.621k i/100ms
   haml 5.0.0.beta.2     2.651k i/100ms
Calculating -------------------------------------
          haml 4.0.7     17.383k (± 3.6%) i/s -     87.534k in   5.042719s
   haml 5.0.0.beta.2     29.101k (± 2.7%) i/s -    145.805k in   5.014494s
Comparison:
   haml 5.0.0.beta.2:    29101.0 i/s
          haml 4.0.7:    17383.1 i/s - 1.67x  slower
```

### after (this branch)
[the result on Travis](https://travis-ci.org/k0kubun/haml_bench/jobs/199978005)

**Now Haml 5 is 3.70x faster than Haml 4.**

```
Rendering: /home/travis/build/k0kubun/haml_bench/templates/slim_bench.haml
Warming up --------------------------------------
          haml 4.0.7     1.698k i/100ms
   haml 5.0.0.beta.2     5.321k i/100ms
Calculating -------------------------------------
          haml 4.0.7     16.892k (± 9.7%) i/s -     84.900k in   5.078316s
   haml 5.0.0.beta.2     62.422k (± 8.4%) i/s -    313.939k in   5.079310s
Comparison:
   haml 5.0.0.beta.2:    62421.8 i/s
          haml 4.0.7:    16892.4 i/s - 3.70x  slower
```

Compared to other template engines ([the same build on Travis](https://travis-ci.org/k0kubun/haml_bench/jobs/199978005)), Haml is not so slow now.

```
Rendering: /home/travis/build/k0kubun/haml_bench/templates/slim_bench.haml
Warming up --------------------------------------
          haml 4.0.7     1.780k i/100ms
   haml 5.0.0.beta.2     5.914k i/100ms
          faml 0.8.1     9.624k i/100ms
        hamlit 2.7.5    12.317k i/100ms
          slim 3.0.7     9.419k i/100ms
Calculating -------------------------------------
          haml 4.0.7     18.322k (± 3.2%) i/s -     92.560k in   5.057934s
   haml 5.0.0.beta.2     62.195k (± 1.2%) i/s -    313.442k in   5.040431s
          faml 0.8.1     99.823k (± 1.5%) i/s -    500.448k in   5.014438s
        hamlit 2.7.5    136.294k (± 2.0%) i/s -    689.752k in   5.062706s
          slim 3.0.7    106.476k (± 1.5%) i/s -    536.883k in   5.043343s
Comparison:
        hamlit 2.7.5:   136293.7 i/s
          slim 3.0.7:   106475.8 i/s - 1.28x  slower
          faml 0.8.1:    99823.3 i/s - 1.37x  slower
   haml 5.0.0.beta.2:    62195.2 i/s - 2.19x  slower
          haml 4.0.7:    18321.7 i/s - 7.44x  slower
```