@k0kubun 2017-02-12
Now Temple's new version was released. Merging.